feat(rename-r3-stage-b1): ribosome.py -> compressor.py + shim (#87)
R3 Stage B.1 per the plan in
~/.claude/plans/ethereal-forging-cookie.md and issue #87.

Renames helix_context/ribosome.py to helix_context/compressor.py so
the canonical filename matches the canonical class (`Compressor`,
established in R3 Stage A). The old path remains as a back-compat
shim that re-exports every module-level name (public + single-
underscore private) from the new location.

Files:
  - helix_context/ribosome.py  -> helix_context/compressor.py (git rename)
  - helix_context/ribosome.py   (new file — shim)

The shim uses a dir()-walk loop to re-export every non-dunder name,
which covers the historical private-name leakage that training and
test files rely on:
  - _parse_json, _EXPRESS_SYSTEM, _splice_system, _PACK_SYSTEM,
    _KV_EXTRACT_SYSTEM, _REPLICATE_SYSTEM
  - Plus all public classes: Compressor, Ribosome (alias), OllamaBackend,
    ClaudeBackend, DeBERTaRibosome, etc.

Verification:
  - Direct import works:        from helix_context.compressor import Compressor
  - Legacy import works:        from helix_context.ribosome import Ribosome
  - Package-level import works: from helix_context import Ribosome
  - All four import paths resolve to the same class object
  - Private name access works:  from helix_context.ribosome import _parse_json
  - tests/test_ribosome.py + test_genome.py + test_retrieval_dimensions.py:
    72/72 pass in 5s
